[EPO-432] Loosen nginx restrictions on websockets.
Before, we had a regexp that was only allowing certain URLs to
be upgraded to websockets.  We also need this for the bokeh urls
that are proxied in the containers.  So I tried just relaxing it
for all URLs, and it seems to work with no regressions I can tell.